[refact] Move hw_model::BusLogger into its own file.
